Do not attempt to delete the Shelf if it hasnt been created yet
git-svn-id: http://svn.wikimedia.org/svnroot/pywikipedia/trunk/pywikipedia@6757 6a7f98fc-eeb0-4dc1-a6e2-c2c589a08aa6

+++ interwiki.py
@@ -536,8 +536,9 @@ class StoredPage(wikipedia.Page):
                '_deletedRevs' ]
                  
     def SPdeleteStore():
-        del StoredPage.SPstore
-        os.unlink(StoredPage.SPpath)
+        if StoredPage.SPpath:
+            del StoredPage.SPstore
+            os.unlink(StoredPage.SPpath)
     SPdeleteStore = staticmethod(SPdeleteStore)
 
     def __init__(self, page):
